SinkC: fix ProbeAckData clear previous buffer with the same address
When the data of the first Probe TtoB are stored in buffer due to !io.release.ready,
the second Probe BtoN (maybe from ProbeHelper) also returns data and thus triggers ProbeAckDataCleaner

Since ProbeAckDataCleaner are meant to clean multiple data from the same req, it should not be triggered between different reqs.
MSHR task sent to SinkC indicates that the last ProbeAckData for this req has already arrived, so its data should be kept.
Any following ProbeAckData, even with the exact same address, should be treated as a new term.